Remove canceled AsyncOperations from channel queues
Canceled {WaitTo}Read/WriteAsync operations currently become nops but still sit in their queue until they're dequeued by a corresponding write or read that's trying to fulfill one of the waiters. This was done intentionally, as it left the code simpler and it was deemed unlikely to be an issue to have these objects sitting around for a bit. But it has since proven disruptive for some applications, which experience a temporary but meaningful "leak". To address that, this PR switches the queues from being array-based to being doubly-linked lists, with cancellation not only marking the operation as canceled but also removing it from the corresponding list.

Doing so required adding a few fields to the operation instance, so compensating changes were also made to try to reduce the size of the instance. This includes propagating a change that had been made to ManualResetValueTaskSourceCore to also apply here: rather than having a field for ExecutionContext and a separate field for a SynchronizationContext/TaskScheduler, use just a single object field and allocate a tuple when both are needed. Needing to store an ExecutionContext is rare, so this is almost always a win. I also utilize some more .NET Core APIs, in particular an overload of CancellationToken.Register that passes in the relevant cancellation token in order to avoid needing to store that same token on the instance. The net result is, at least on .NET Core, with this change the AsyncOperation instance actually gets a bit smaller than it was previously.

As new null checks were being added, I also switched the whole project over to using is / is not rather than == / !=.

As part of this change, to help flow a bit of state about the purpose of each instance, I refactored the AsyncOperation type in order to have concrete types per scenario (e.g. ReadAsync vs WriteAsync vs WaitToReadAsync vs WaitToWriteAsync) and also pushed a bunch of members from the generic down to a non-generic base, which should help a bit further with throughput but also Native AOT size.

All in all, this should address the memory issue while also cleaning up the code a bit and having a small positive impact on overall performance. Running our microbenchmarks locally shows things to be net neutral or slightly beneficial.