Attempt to fix the rare stack overflow tests issue
Our CI hits a strange issue in the stack overflow tests about once a day. We are hitting
another stack overflow on the thread that handles the stack overflow. One theory I have is
that an activation signal kicks in and somehow results in overflowing the stack overflow
handling helper stack. This change disables activation signals for the current thread
once stack overflow is hit.
We'll see if that fixes the problem.