DriveInfo.Linux: use procfs mountinfo for formats and mount point paths.
For the format, mountinfo includes a string 'filesystem type' field which is more granular than the statfs f_type numeric field.
For example: the ext2, ext3 and ext4 are distinct mountinfo values which have the same same f_type value.

For the mount paths, mountinfo provides the mount point paths relative to the process's root directory.
When the process runs with changed root dir (chroot), the paths are adjusted for it.