Fix interpreter double and 64 bit constant processing
There is a bug in extracting 64 bit and double constants on both the
interpreter compiler and execution sides. The problem was that the lower
32 bits were extracted as signed 32 bit int and then casted to signed 64
bit int and or-ed to the other 32 bits shifted up. The problem is that
when the lower 32 bits were signed, the sign got extended to the upper
32 bits and the combined result was incorrect.

This fixes 24 codegen bringup tests with the interpreter that were
previously failing with "Assert.Equal() Failure: Values differ".