Arm64 SVE: Do not remove embedded conditional selects
Fixes #116140

When an embedded SVE instrinsic is created, during lowering a conditional select is created to embed the operation. This required to ensure there is a mask for the embedded intrinsic to use.

Ensure that the embedding conditional select can only be removed if the embedded intrinsic is also removed.