JIT: Use flowgraph annotations to scale loop blocks in `optSetBlockWeights` (#116120)
Part of #107749. This replaces optMarkLoopHeads and optFindAndScaleLoopBlocks with the graph-based loop visitor pattern we now use elsewhere in the JIT. I tried to preserve the heuristics in optScaleLoopBlocks, but a side effect of this change is we no longer scale unnatural loops. However, we no longer rely on lexical loop finding, so we gain some precision as well.

Once this and graph-based loop inversion have been merged, we will no longer have any fgRenumberBlocks calls.