[mono][interp] Remove invalid addition of conversion opcodes (#116130)
As part of a conditional branch we were adding an implicit conversion opcode (if needed) after `td->last_ins`. If the branch is the first opcode in the basic block then this actually adds an instruction in another basic block, which is clearly incorrect. This seems to have been done like this early in the implementation, in order to achieve certain il offset mapping of instructions, but everything around this area has been refactored since, so it shouldn't be necessary anymore.

With tiering disabled, this was crashing b69225 il test.