Disable CoreCLR interpreter test for JIT stress (#116219)
There is a problem in the interpreter when handling calls to code that
doesn't have native code yet, but that has precode already. The call to
MethodDesc::PrepareInitialCode triggers an assert down its call chain
due to the prestub existence.

This change disables running the interpreter test with JIT stress until
the problem is solved.