Fix LDTOKEN in interpreter
The CEE_LDTOKEN compilation incorrectly assumes that the resolved token
contains either hClass or hField or hMethod and it checks for the hClass
first. But the hClass is present always, so resolving token for method
or field doesn't work correctly.

This change fixes it by reordering the checks in an order matching what JIT
does.