Share allocation helpers between CoreCLR and NativeAOT (#115339)
Share assembly version of allocation helpers between CoreCLR and NativeAOT. This allows better code reuse and removes the last usage of helper method frames in the CoreCLR runtime, cleanup is expected to be done separetely.

The slow paths on CoreCLR differ from NativeAOT in few details. They don't expect the GC_ALLOC_FINALIZE, GC_ALLOC_ALIGN8 and GC_ALLOC_ALIGN8_BIAS flags to be passed the correct values and if they are not present then they get inferred from the method table. This means that helpers like RhpNewFinalizable stay exclusive to NativeAOT and remain unused on CoreCLR. This is done intentionally so we can have a single slow path (RhpNew for objects, RhpNewArray for arrays) that can be assigned to any JIT helper. The slow path is used when allocation tracking is enabled or for other diagnostic options.

CoreCLR stores the per-thread allocation context in separate TLS variable, while NativeAOT track it as part of the main per-thread information. We abstract it by introducing INLINE_GET_ALLOC_CONTEXT macro that returns the pointer to ee_alloc_context structure.

FastAllocateString helper in CoreCLR has a changed signature to take the string method table as the first parameter. This mimics the same idea as used by NativeAOT where we can use slow array allocation helper and fast string allocation helper interchangably.

Platform specific notes:

- linux-x86 didn't have the helpers implemented; the implementation is added and passed some smoke tests.
- linux-riscv64 implementation was heavily rewritten. The main change is to establish function frame and move arguments into non-volatile registers. We support platform versions without TLS descriptor and this ensures that we only save minimum number of registers to stack to enable INLINE_GET_ALLOC_CONTEXT doing function call that trashes volatile registers.
- linux-arm helpers for Align8/Misalign cases were enhanced to include padding with "free object" in the fast path.

Contributes to #95695
---------

Co-authored-by: Tomasz Sowiński <tomeksowi@gmail.com>