Add `--arch` option for `--list-runtimes` and `--list-sdks` in host (#116078)
Allow specifying --arch <arch> for --list-runtimes and --list-sdks commands in host. Must be in the order --list-runtimes/sdks --arch <arch>.
- If not specified, behaviour remains the same - list runtimes/SDKs for the invoked dotnet (for SDKs, any paths specified any found global.json are respected).
- If specified and not the same as the host architecture, searches for registered or default install locations for that architecture and lists runtimes/SDKs found there.
- If specified and the same as the host architecture, behaves the same as if not specified - note that this means it will not search for registered/default installs of the host architecture and only list runtimes/SDKs relative to the invoked dotnet.

New tests are in a HostCommands test class. This also moves some existing tests for host commands into the new class. This does remove tests running the host commands with the multi-level lookup environment variable set, but I don't think those are interesting anymore - it has been disabled since 7.0 and should always be disabled (unlike for actually running an app, where it can be based on the tfm).