Make the EECodeGenManager API surface narrower. Rename some locals removing critical section notion and replace with lock.
Rework the CodeHeapIterator.
Rename members to more accurately reflect their purpose.
Generally move to more canonical C++.

EECodeGenManager now has a iterator counter. This is
used to track the number of outstanding iterators and
can be used to defer deletes. Adds are still valid since
the iterator will only operate on the state at the time the
iterator was created.