Do not return from CTRL_SHUTDOWN_EVENT
Windows will kill the process after it completes handling CTRL_SHUTDOWN_EVENT

To avoid this, we add a delay to our handler.  This allows the process
to exit before the handler returns.

We use HostOptions.ShutdownTimeout for the delay.  This does mean that
other SIGTERM handlers will not run, but that's a tradeoff we have to
make to ensure the process isn't killed.