Delete MultilevelSDKLookup test class (#116396)
Multi-level lookup was disabled in 7.0 - for SDK lookup (unlike runtime lookup, which is affected by tfm), it is always disabled. I don't think tests validating that it is indeed disabled regardless of if/how the environment variable is set are still useful. Aside from SdkResolutionError, all the cases in MultilevelSDKLookup are covered in SDKLookup. This change moves that one test case and deletes the MultilevelSDKLookup test class.