Fix stresslog facility and log level mix-up. (#116467)
As part of commit:

https://github.com/dotnet/runtime/commit/6bd555eefd41282282da4c5dfe28b8600946b744

some stress log macro implementations got changed, but the call to
StressLog::LogMsg in gcenv.h and stresslog.h got level and facility arguments
switched, leading to no logging when these macros are used.

Fix restore the order of the level and facility arguments in call to
StressLog::LogMsg.