Fix BulkType event for GC static bases (#116479)
.NET Native had extra code to handle this but it ended up under a ProjectN ifdef and that (unsurprisingly) got deleted:

https://github.com/dotnet/corert/blob/c2cc114e704bcf7a6af57d67b1e9e92d85eea1d2/src/Native/Runtime/rheventtrace.cpp#L200-L218

The `We will not be able to get the osModuleHandle for these` comment is more recent (#94673) but that approach will not work because event pipe consumers won't know what to do with these.

So restore .NET Native behavior (in spirit - we deleted most of the code the .NET native implementation relied on).