Map version for Tahoe compatibility.
macOS Tahoe returns a compatibility version, 16, for macOS 26 unless it is built with Xcode 26's SDK.
As we did with Big Sur, this maps the compatibility version 16 to 26.

The intention is that we will be on the new SDK by the time macOS 27 rolls out. If not, then we will need to add another compatibility map, most likely.
It does not appear that iOS, tvOS, or Catalyst return compatibility numbers, so they are excluded from doing any mapping.