Harden `Ping_TimedOut_*` tests (#116500)
Depending on the network configuration, we can get an ICMP with `Type=Destination Unreachable, Code=Network Unreachable` from middlewares when pinging an "unreachable" IP. In #115358 we see this happening on Helix Mac machines, but I was able to reproduce it on an office LAN-connected Windows PC too.

This PR attempts to fix address this problem by repeating the ping tests on other "types" of unreachable hosts until we actually catch a timeout. (We could instead just skip the test on the first DestinationNetworkUnreachable, but then we would loose some functional coverage.)

Marking as draft until we get a few successful outerloop runs.