Arm64 SVE: Do not remove embedded conditional selects (#116144)
* Arm64 SVE: Do not remove embedded conditional selects

Fixes #116140

When an embedded SVE instrinsic is created, during lowering a conditional select is created to embed the operation. This required to ensure there is a mask for the embedded intrinsic to use.

Ensure that the embedding conditional select can only be removed if the embedded intrinsic is also removed.

* Check for EmbeddingMaskOp in LowerHWIntrinsicCndSel

* Fix HasDisasmCheck

* fix formatting

* debug and comment nits

* check all cndsel ops in OperMayThrow

* Add OperOrEmbeddedChildrenMayThrow()

* Replace with recursive OperOrContainedChildrenMayThrow()

* remove pointless assert

* rename to NodeOrContainedOperandsMayThrow

* use for loop instead of walker

* Update src/coreclr/jit/lowerarmarch.cpp

* add test env

* update HasDisasmCheck

---------

Co-authored-by: Kunal Pathak <Kunal.Pathak@microsoft.com>