Add ability to filter interpreted methods by assembly (#116569)
* Add ability to filter interpreted methods by assembly

This change adds support for assembly name to the methodset stuff so
that it can be used to select interpreted methods or methods to generate
interpreter compilation dump for.

To preserve parity, I've made the same change to the JIT's version of
the same functionality.

This change is a prerequisity for an upcoming change that enables
running coreclr tests interpreted.

* PR feedback - added docs and fixed indentation