Improve handling of dynamic types in BulkType events
Dynamically allocated types (e.g. result of `MakeGenericType`) don't have debug information associated with them so we can't really reconstruct their names from PDB. However, we know the canonical form of the type and that one does have debug info. Emit RVA of the canonical form instead a non-sensical number.

Consumers can still distinguish this is a different thing using the `TypeID` field (that is still the MethodTable pointer). The canonical form is only used for `TypeNameID`, which is RVA in native AOT case and definition token in the JIT case. The consumer can quickly see that `ModuleID + `TypeNameID != TypeID`.

We could generate enough information to fully reconstruct the type name because the event can actually carry composition information, but that support was deleted in #94673 on native AOT side and it's currently not consumed by any tools I know of. It can be done later; Project N went to retirement with not even having the canonical forms in traces.