Generate module rundown events (#116338)
* Generate module rundown events

This is sufficient to get module information in `dotnet-gcdump collect`. The tool itself will need tweaks because it only looks at IL paths right now, but this should be all that's necessary to build on top.

* Put back GUID

* Fix linux

* More Linux fixing

* linux

* wow

* Feedback

* Maybe it works?

* Apply suggestions from code review

Co-authored-by: Jan Kotas <jkotas@microsoft.com>

* Build break

* Copilot commit

Prompt: The use of mbstowcs_s is incorrect here. The input buffer is UTF-8, output is UTF-16. Replace the incorrect use of mbstowcs_s with MultiByteToWideChar.

* Update issues.targets

---------

Co-authored-by: Jan Kotas <jkotas@microsoft.com>