Fix stackwalk from interpreted EH funclet (#116640)
* Fix stackwalk from interpreted EH funclet

The stack walking for GC is broken when an interpreted EH funclet was
on the stack. The reason is that call to the interpreted funclet doesn't
have a transition frame and that case was not handled properly. Also,
I have found that the ExInfo::m_csfEHClause was not being set for the
interpreted clauses at all.

This change fixes it by using the InterpreterFrame address as the
m_csfEHClause (so that it is above all the interpreter frames) and
fixing the `ExInfo::FindParentStackFrameHelper` to recognize the special
instruction pointer value that we get as a parent IP for the first
interpreted frame under an InterpreterFrame.

That special value has been zero and I have added a named constant for
it so that it is clear at the points where it is being used.

* PR feedback

* Update src/coreclr/vm/exceptionhandling.cpp

Update comment

Co-authored-by: Jan Kotas <jkotas@microsoft.com>

---------

Co-authored-by: Jan Kotas <jkotas@microsoft.com>