Update host test command execution logging: include PID on exit, remove confusing OK/FAIL, add caller name (#116658)
When launching a process to test, host tests write to the console on launch, wait for exit, and exit. Currently, they also log OK/FAIL based on the exit code and a parameter for whether failure is expected. All tests already do their own validation for if the command should pass/fail - having this extra logging is not useful and can be actively confusing, since it relies on tests also passing in that parameter (even though that is not used for actual validation/asserts).

Update the logging to remove the OK/FAIL, include the PID on exit, and add the caller name (for test identification).