DriveInfo.Linux: use procfs mountinfo for formats and mount point paths. (#116102)
* DriveInfo.Linux: use procfs mountinfo for formats and mount point paths.

For the format, mountinfo includes a string 'filesystem type' field which is more granular than the statfs f_type numeric field.
For example: the ext2, ext3 and ext4 are distinct mountinfo values which have the same same f_type value.

For the mount paths, mountinfo provides the mount point paths relative to the process's root directory.
When the process runs with changed root dir (chroot), the paths are adjusted for it.

* Remove stackalloced buffer.

* Resolve symbolic links.

* Fix Common.Tests build on non-Linux.

* Add test that gets the DriveFormat for /tmp.

* Use test to debug CI failure.

* Fix check for root mount path.

* pal_mount: use buffer length based on API used.

* Fix not calling native GetFileSystemTypeNameForMountPoint on non-Linux.

* Add back using 'f_type' when 'f_fstypename' is not available.

* UnixFileSystemTypes: prefer Linux file system type strings as names, and remove duplicate value members.

* Fix wasi build breakage.

---------

Co-authored-by: David Cantú <dacantu@microsoft.com>