Add few interpreter bringup diagnostic improvements (#116699)
This change improves the diagnosability of the interpreter issues during
the bringup. It prints out unsupported IL opcodes when we hit one and
also makes the assert go through the assertAbort instead of the regular
assert, since the regular one pops out a dialog, so it is bad for batch
testing.