JIT: extend indirect virtual stub optimization to arrays
Add support for array interface methods invoked via indirect virtual stubs.
Trim out a layer of temps in the importer. Remember the base method handle
and use that for detecting the GetEnumerator pattern and for array interface
method devirtualization. Mark the  SZGenericArrayEnumerator ctor as aggressively
inlined; when inlined we can often prove the enumerator (conditionally) doesn't
escape.

Part of #108913