Fix DAC-ification of the InterpreterFrame::SetContextToInterpMethodContextFrame (#116797)
There is an incorrect cast of `this` to TADDR that was resulting in the
DAC side address being stored in the context. That was leading to an
assert in the stack walking code on the DAC side that was crashing
WinDbg.

This change replaces C-style cast with the proper dac_cast.