Fix unhandled exception in tiered compilation thread (#116716)
* Fix unhandled exception in tiered compilation thread

The detection the unhandled exception is not working in the case of
the tiered compilation thread, because there is no indication that
the exception is going to be caught in the `CEEInfo::runWithErrorTrap`.
So the EH assumes the exception is going to be unhandled and
exits the process.
The unhandled exception message was not seen in the repro case
of the issue likely because the tiered compilation thread console
output didn't get flushed. When running under a debugger and stepping
over the `InternalUnhandledExceptionFilter_Worker`, it was actually
printed out.

The fix is to add `DebuggerU2MCatchHandlerFrame` to the
`CEEInfo::runWithErrorTrap`. That allows the EH to see that the
exception will be caught there and let it flow through the
CallDescrWorker to reach that trap.

Closes #116676

* Add regression test and fix GC mode

* Make the regression test actually work

It needs to be compiled as optimized and the Console.WriteLine is needed to repro
the issue.

* Modify the regression test to not to print to console

* PR feedback

* Update src/coreclr/vm/jitinterface.cpp

Fix contract

Co-authored-by: Jan Kotas <jkotas@microsoft.com>

---------

Co-authored-by: Jan Kotas <jkotas@microsoft.com>