Reflection.Emit should respect RVAs of 0 (#116766)
The underlying Reflection.Emit infrastructure was treating an RVA of 0 as "normal" and indexing into the dynamic IL data stream. This caused confusion in the rest of the system since an RVA of 0 means there is no IL body (that is, no IL Header). Making Reflection.Emit consistent ensures that scenarios relying on an RVA of 0 now operate identical with those emitted by .NET compilers.

This change also removes some of the checks that were put in place for Reflection.Emit and relies on the EE at runtime to handle those cases.