Ldvirtftn implementation for interpreter (#116661)
Implement more complex call scenarios
- ldvirtftn based (generic virtual calls)
- code pointer based (certain static virtual calls)
- ldvirtftn path to delegate creation

And as a bonus, to keep the tests working...

Rework where we attach intepreted code to a method, so that we never put the interpreter code bytes into the vtable (with this change, the vtable, etc, always have the InterpreterPrecode in it.

Add scheme for enabling logging of the CallStubGenerator.

Fix calling convention handling for jit calling convention to interpreter calling convention for Unix X64 and Windows Arm64.