Add trimming intrinsic for EventInfo.EventHandlerType.GetMethod("Invoke") (#116931)
Because ILLink and ILC keep the `Invoke` method on delegates, there
shouldn't be a warning for the common pattern of
`EventInfo.EventHandlerType.GetMethod("Invoke")`.

It's possible to break this logic by using a custom `EventInfo`, but
we don't support the scenario where this breaks the contract that
`EventHandlerType` should be a delegate.