JIT: GDV Cleanup Phase
If a method has GDVs and the JIT is able to to improve types during inlining,
run a pass after inlining that tries to resolve GDVs using those improved types.

This typically happens when the object tested in a GDV is returned from some kind of
factory method where the exact type becomes apparent only after inlining.

Contributes to #25448.