Fix absolute path check when loading hostfxr/hostpolicy/coreclr (#116776)
On Windows, we were incorrectly only checking for <letter>: and incorrectly determining that UNC and device paths were not absolute. This updates pal::is_path_rooted to match the managed Path.IsPathRooted and adds a pal::is_path_absolute with the logic of Path.IsPartiallyQualified

This also adds an error message when we can't resolve hostfxr and tests for device paths.