Do not assert IsSupported in DefaultValueAttribute ctor
We have defined behaviors for this. We have tests that test the defined behaviors. I guess I'm the first person to run those tests with a Debug version of CoreLib and hit the assert. (Libraries testing typically uses Release runtime but the only time I build Release runtime is for perf testing.)