Enable support for InlinedCallFrame in the interpreter
This change enables using InlinedCallFrame around pinvokes called from
the interpreter. It ensures that the InlinedCallFrame caller can be set
to the interpreted code and that stack walk from the InlinedCallFrame
through the interpreted code and then through the InterpreterFrame works
properly.

As an example and to test it, I have added an InlinedCallFrame to the temporary
implementation of the INTOP_GC_COLLECT and changed it to call a QCALL.