JIT: Update switch likelihoods after peeling dominant case (#116984)
Fixes #116962. After peeling a switch block such that the switch is dominated by a conditional check for the dominant case, set the dominant switch case's edge likelihood to zero, and redistribute this extra likelihood among the other switch successor edges. This also replaces BBswtDesc::bbsDominantFraction with the edge likelihood of the dominant case, as the latter value is maintained through flow updates.