Shared generics opcodes (#116990)
Implement the vast majority of the opcodes that are affected by shared generics

box, unbox, unbox.any, castclass, isinst, ldtoken, newarr, newobj of an mdarray
Add logic to enhance pretty printing of the Interp opcodes for these things. We can now see strings, and various magic pointer values get displayed in the disasm outputs.

Handle the 0 indirections generic lookup scenario

Many opcodes were able to use a set of generic helpercall opcodes. These opcodes have a naming convention of

a stands for addr of dvar or svar
g stands for do a generic lookup
p stands for pointer (acquired from either an svar or a dataitem)
v stands for a void return.
Its not fully descriptive, but naming is hard, and this works.