Http3RequestStream: read EOS after trailers (#117026)
This fixes the primary bug captured in #60118 without attempting to drain the stream and without touching the Dispose logic.

We expect trailers to be sent in a single HEADER frame, which is read by `ReadHeadersAsync` but that method doesn't read the EOS. Despite of that, we set `_responseDataPayloadRemaining = -1` afterwards

https://github.com/dotnet/runtime/blob/27c8fe04453195039686fb375a73fac7ff1ea968/src/libraries/System.Net.Http/src/System/Net/Http/SocketsHttpHandler/Http3RequestStream.cs#L577-L583

which then makes `Http3RequestStream` pretend that it reached EOS in subsequent read calls without ever trying to issue the read to the underlying `QuicStream`:

https://github.com/dotnet/runtime/blob/0d628dae1cb5b8d0eb76d9549134edfe30d6219c/src/libraries/System.Net.Http/src/System/Net/Http/SocketsHttpHandler/Http3RequestStream.cs#L1340-L1344

This can be easily fixed by issuing one more read after reading the headers.