Replace `CordbProcess::GetSharedDomain` with `GetAppDomain` (#117037)
This removes the 'shared domain' idea (which was basically an empty CordbAppDomain) from the ICorDebug* implementations. Everything operates on the (one and only) AppDomain. There's plenty more that could be cleaned up with app domains here (around there only being one now), but what's in this change seemed like a reasonably contained step.